+++ IzvorniKod/backend/src/main/java/com/example/demo/rest/StudentController.java
@@ -1,10 +1,7 @@
 package com.example.demo.rest;
 
-import com.example.demo.users.Student;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @RestController
 @RequestMapping(path = "/student")
